Pairing: Links resources (from resource)
- wrong resource(s)

---
Comment after the pairing session

The requirement was to display link metadata in a link#show page.
Show page requires an params[:id] to fetch the data from the
database. The params[:id] was missing because in the routes file
there was `resource :links` instead of `resources :links`.

The change broke a redirections_spec test but there was no time
to look at it.